refactor(creation): Extract AdeptPowersCard modal to subfolder
- Create /components/creation/adept-powers/ directory
- Extract inline modal JSX to AdeptPowerModal.tsx component
- Use BaseModalRoot for proper accessibility (focus trapping, ESC key)
- Modal manages its own internal state (search, selection, level)
- Parent controls isOpen and receives onAdd callbacks
- Reduce AdeptPowersCard from 849 to 586 lines (31% reduction)

Fixes #235
