+++ components/creation/adept-powers/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Adept Powers Components
+ *
+ * Modal-based adept power selection for character creation.
+ */
+
+export { AdeptPowerModal, type AdeptPowerModalProps } from "./AdeptPowerModal";
